Add new generic component to inject variable in input/textarea in scenes
